+++ src/main/java/org/opentripplanner/graph_builder/module/GtfsModule.java
@@ -58,27 +58,31 @@ public class GtfsModule implements GraphBuilderModule {
    */
   private final ServiceDateInterval transitPeriodLimit;
   private final List<GtfsBundle> gtfsBundles;
+  private final FareServiceFactory fareServiceFactory;
+  private final boolean discardMinTransferTimes;
   private DataImportIssueStore issueStore;
-  private FareServiceFactory fareServiceFactory;
   private int nextAgencyId = 1; // used for generating agency IDs to resolve ID conflicts
 
-  public GtfsModule(List<GtfsBundle> bundles, ServiceDateInterval transitPeriodLimit) {
+  public GtfsModule(
+    List<GtfsBundle> bundles,
+    ServiceDateInterval transitPeriodLimit,
+    FareServiceFactory fareServiceFactory,
+    boolean discardMinTransferTimes
+  ) {
     this.gtfsBundles = bundles;
     this.transitPeriodLimit = transitPeriodLimit;
+    this.fareServiceFactory = fareServiceFactory;
+    this.discardMinTransferTimes = discardMinTransferTimes;
+  }
+
+  public GtfsModule(List<GtfsBundle> bundles, ServiceDateInterval transitPeriodLimit) {
+    this(bundles, transitPeriodLimit, null, false);
   }
 
   public List<String> provides() {
     List<String> result = new ArrayList<>();
     result.add("transit");
     return result;
-  }
-
-  public List<String> getPrerequisites() {
-    return Collections.emptyList();
-  }
-
-  public void setFareServiceFactory(FareServiceFactory factory) {
-    fareServiceFactory = factory;
   }
 
   @Override
@@ -106,6 +110,7 @@ public class GtfsModule implements GraphBuilderModule {
         GTFSToOtpTransitServiceMapper mapper = new GTFSToOtpTransitServiceMapper(
           gtfsBundle.getFeedId().getId(),
           issueStore,
+          discardMinTransferTimes,
           gtfsDao
         );
         mapper.mapStopTripAndRouteDatantoBuilder();
@@ -155,14 +160,14 @@ public class GtfsModule implements GraphBuilderModule {
     graph.calculateTransitCenter();
   }
 
-  /* Private Methods */
-
   @Override
   public void checkInputs() {
     for (GtfsBundle bundle : gtfsBundles) {
       bundle.checkInputs();
     }
   }
+
+  /* Private Methods */
 
   /**
    * This method have side-effects, the {@code stopTimesByTrip} is updated.