+++ docs/Changelog.md
@@ -61,6 +61,8 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - Transit model - Part 3 [#4176](https://github.com/opentripplanner/OpenTripPlanner/pull/4176)
 - Add Siri realtime occupancy info [#4180](https://github.com/opentripplanner/OpenTripPlanner/pull/4180)
 - Add gtfs stop description translations [#4158](https://github.com/opentripplanner/OpenTripPlanner/pull/4158)
+- Upgrade graphql-java to version 18 [#4194](https://github.com/opentripplanner/OpenTripPlanner/pull/4194)
+- Add option to discard min transfer times [#4195](https://github.com/opentripplanner/OpenTripPlanner/pull/4195)
   [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)
 
 ## 2.1.0 (2022-03-17)